C automatically typecast the return value of malloc
C automatically promotes pointers to void to any type on assignment.
So do not need to typecast the return value of malloc() to the lvalue's
type used in assignments.

	char *p;

	/* give me 2kb */
	p = malloc(2048);
	if (!p)
		perror("malloc");

or allocate a struct:
	struct treasure_map *map;

	map = malloc(sizeof(struct treasure_map));
	if (!map)
		perror("malloc");

The C++ programming language, however, does not perform automatic void
pointer promotin. Consequently, user of C++ need to typecast malloc()'s
return value as follows:

	char *name;

	/* allocates 512 bytes */
	name = (char *)malloc(512);
	if (!name)
		perror("malloc");

Change-Id: Ie5fdb5f5e08913a879a8bb9887f7160838b5bf01